# [TIR][Schedule] Add annotate_buffer_access primitive